Finished removal of Storage and DocumentMap - createRef, EntityId, and getEntityId were moved to create-ref.ts - StorageManager added syncCell and close was added to IStorageManager interface.
