+++ packages/charm/src/iterate.ts
@@ -1,16 +1,15 @@
 import { isObject, Mutable } from "@commontools/utils/types";
 import {
-  Cell,
+  type Cell,
   createJsonSchema,
-  type IExtendedStorageTransaction,
   isCell,
   isStream,
-  JSONSchema,
-  JSONSchemaMutable,
+  type JSONSchema,
+  type JSONSchemaMutable,
   type MemorySpace,
-  RecipeMeta,
+  type RecipeMeta,
   type Runtime,
-  RuntimeProgram,
+  type RuntimeProgram,
 } from "@commontools/runner";
 import { Charm, CharmManager, charmSourceCellSchema } from "./manager.ts";
 import { buildFullRecipe, getIframeRecipe } from "./iframe/recipe.ts";
