+++ packages/runner/src/data-updating.ts
@@ -1,7 +1,7 @@
 import { isRecord } from "@commontools/utils/types";
 import { getLogger } from "@commontools/utils/logger";
 import { ID, ID_FIELD, type JSONSchema } from "./builder/types.ts";
-import { createRef } from "./doc-map.ts";
+import { createRef } from "./create-ref.ts";
 import { isCell, RegularCell } from "./cell.ts";
 import { resolveLink } from "./link-resolution.ts";
 import {
